clippy updates

+++ api/cmd/server.rs
@@ -13,7 +13,7 @@ pub async fn main(args: Args) -> Result<(), crate::error::Error> {
     let shutdown = GracefulShutdown::new();
     let config = crate::server::Config {
         bind_address: Some(envoption::with_default("BIND_ADDRESS", "127.0.0.1")?),
-        bind_port: envoption::with_default("BIND_PORT", 6543 as u16)?,
+        bind_port: envoption::with_default("BIND_PORT", 6543_u16)?,
         database: database_configuration_from_env()?,
         redis_url: None,
         redis_queue_prefix: None,
